Huge parser improvement: implemented netlist hierarchy generation. This not only completes the mid-term planned parser features, but also allows the user to create his own custom libraries simply by creating an arbitrary symbol and a schematic of the same filename.
To be able to use the created library, the user afterwards must add the
folder containing the files to the list of libraries in the settings. In
the future this will also be automatic by considering the current file
folder as an additional library.

This feature completes the schematic capture and netlist generation
planned (and programmed) for the last couple of years!!!