HUGE memory leaks fix.
Fixed memory leaks in documents open/close. Previously the destructors were not
defined, leaving dangling pointers. This was not previously noticed, however
when performing big simulations the unused data of closed documents make a big
difference. For example, previously if a raw simulation of about 100 MB was opened
Caneda would use about 230 MB of memory (a dangling pointer during file loading was
also lost, doubling the memory size). After the document was closed the same amount
of memory was occupied. If afterwards more documents were opened, the lost memory
would accumulate, easily creating problems for the user. With this fix, if the same
document of 100 MB is opened, only 130 MB are used (~100 MB for the file and 30 MB
for the rest of the application). After the document is closed, the 100 MB of the
file are freed, leaving only the 30 MB of the rest of the application. In this way,
one could open multiple big files without having any memory problem.